# apps/web/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "properties": {
          "code": {
            "enum": [
              "not_found",
              "revision_conflict",
              "candidate_limit",
              "adopted_candidate",
              "candidate_archived",
              "candidate_provenance_immutable",
              "project_task_locked",
              "project_group_conflict",
              "protected_project",
              "project_has_successors",
              "data_integrity_error",
              "validation_error",
              "runtime_unavailable"
            ],
            "title": "Code",
            "type": "string"
          },
          "current_candidate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Candidate"
              },
              {
                "type": "null"
              }
            ]
          },
          "field_errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Field Errors",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "FieldError": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "ModelPackagePredictorStatus": {
        "properties": {
          "predictive_family": {
            "title": "Predictive Family",
            "type": "string"
          },
          "runtime_type": {
            "title": "Runtime Type",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "target",
          "runtime_type",
          "predictive_family"
        ],
        "title": "ModelPackagePredictorStatus",
        "type": "object"
      },
...
      "ModelPackageStatus": {
        "properties": {
          "active_runtimes": {
            "items": {
              "type": "string"
            },
            "title": "Active Runtimes",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "manifest_sha256": {
            "title": "Manifest Sha256",
            "type": "string"
          },
          "predictors": {
            "items": {
              "$ref": "#/components/schemas/ModelPackagePredictorStatus"
            },
            "title": "Predictors",
            "type": "array"
          },
          "quality_report": {
            "$ref": "#/components/schemas/ModelQualityReport"
          },
          "supported_runtimes": {
            "items": {
              "$ref": "#/components/schemas/RuntimeAvailability"
            },
            "title": "Supported Runtimes",
            "type": "array"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "version",
          "task_id",
          "manifest_sha256",
          "active_runtimes",
          "supported_runtimes",
          "predictors",
          "quality_report"
        ],
        "title": "ModelPackageStatus",
        "type": "object"
      },
      "ModelQualityReport": {
        "additionalProperties": true,
        "properties": {
          "split": {
            "title": "Split",
            "type": "string"
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/ModelQualityTarget"
            },
            "title": "Targets",
            "type": "array"
          }
        },
        "required": [
          "split",
          "targets"
        ],
        "title": "ModelQualityReport",
        "type": "object"
      },
...
      "ModelTrainingDataPage": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/TrainingDataColumn"
            },
            "title": "Columns",
            "type": "array"
          },
          "curation_summary": {
            "$ref": "#/components/schemas/ModelTrainingCurationSummary"
          },
          "feature_dataset_digest": {
            "title": "Feature Dataset Digest",
            "type": "string"
          },
          "feature_pipeline_id": {
            "title": "Feature Pipeline Id",
            "type": "string"
          },
          "feature_pipeline_version": {
            "title": "Feature Pipeline Version",
            "type": "string"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "parent_conditions": {
            "title": "Parent Conditions",
            "type": "integer"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/TrainingDataRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "source_data_digest": {
            "title": "Source Data Digest",
            "type": "string"
          },
          "stage": {
            "enum": [
              "curation",
              "selected",
              "features"
            ],
            "title": "Stage",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          },
          "target_label": {
            "title": "Target Label",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "training_unit": {
            "enum": [
              "individual_observation",
              "parent_condition_mean"
            ],
            "title": "Training Unit",
            "type": "string"
          }
        },
        "required": [
          "stage",
          "target",
          "target_label",
          "source_data_digest",
          "feature_dataset_digest",
          "feature_pipeline_id",
          "feature_pipeline_version",
          "training_unit",
          "total",
          "parent_conditions",
          "offset",
          "limit",
          "columns",
          "rows",
          "curation_summary"
        ],
        "title": "ModelTrainingDataPage",
        "type": "object"
      },
...
      "ResolvedTaskDefinition": {
        "additionalProperties": false,
        "properties": {
          "application": {
            "$ref": "#/components/schemas/ApplicationCapability",
            "default": {
              "candidate_excel_export": false,
              "candidate_excel_import": false,
              "project_creation": true
            }
          },
          "data_explorer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataExplorerCapability"
              },
              {
                "type": "null"
              }
            ]
          },
          "runtime_capability": {
            "$ref": "#/components/schemas/RuntimeCapability"
          },
          "task_definition": {
            "$ref": "#/components/schemas/TaskDefinition"
          }
        },
        "required": [
          "task_definition",
          "runtime_capability"
        ],
        "title": "ResolvedTaskDefinition",
        "type": "object"
      },
...
      "RuntimeAvailability": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "runtime_type": {
            "title": "Runtime Type",
            "type": "string"
          }
        },
        "required": [
          "runtime_type",
          "available"
        ],
        "title": "RuntimeAvailability",
        "type": "object"
      },
...
      "TaskCatalogItem": {
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/ResolvedTaskDefinition"
          },
          "starter_candidate": {
            "$ref": "#/components/schemas/CandidateInput"
          }
        },
        "required": [
          "definition",
          "starter_candidate"
        ],
        "title": "TaskCatalogItem",
        "type": "object"
      },
...
      "TrainingDataColumn": {
        "properties": {
          "group": {
            "enum": [
              "識別",
              "原値",
              "正規化",
              "判定",
              "入力",
              "特徴量",
              "実測"
            ],
            "title": "Group",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "required": [
          "key",
          "label",
          "group"
        ],
        "title": "TrainingDataColumn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Material Decision Workbench API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects/{project_id}/model-package": {
      "get": {
        "operationId": "getProjectModelPackage",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelPackageStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Model Package"
      }
    },
    "/api/projects/{project_id}/model-package/training-data": {
      "get": {
        "operationId": "getProjectModelTrainingData",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "stage",
            "required": false,
            "schema": {
              "default": "selected",
              "enum": [
                "curation",
                "selected",
                "features"
              ],
              "title": "Stage",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelTrainingDataPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Model Training Data"
      }
    },
...
    "/api/projects/{project_id}/task-definition": {
      "get": {
        "operationId": "getProjectTaskDefinition",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedTaskDefinition"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Task Definition"
      }
    },
...
    "/api/task-definitions": {
      "get": {
        "operationId": "listTaskDefinitions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskCatalogItem"
                  },
                  "title": "Response Listtaskdefinitions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Task Definitions"
      }
    }
  }
}